+++ src/routes/blog/[slug].svelte
@@ -8,53 +8,12 @@
     import loadingGif from "$lib/assets/loading.gif";
     import SEO from "$lib/components/SEO.svelte";
     import NoBlog from "$lib/components/blog/NoBlog.svelte";
+    import MarkDown from "$lib/components/MarkDown.svelte";
 
     let post;
     let loading = true;
     let title = "Loading...";
     let content = "Loading...";
-
-    const md = new MarkdownIt({
-        html: true,
-        linkify: true,
-        breaks: true,
-        typographer: true,
-        langPrefix: "hljs-",
-        highlight: function (str, lang, attrRaw) {
-            const attrs = attrRaw.split(/\s+/g);
-            const lineNumbers = attrs.includes("ln");
-
-            let code =
-                lang && hljs.getLanguage(lang)
-                    ? hljs.highlight(str, {
-                          language: lang,
-                          ignoreIllegals: true,
-                      }).value
-                    : md.utils.escapeHtml(str);
-
-            if (lineNumbers) {
-                code = applyLineNumbers(code);
-            }
-
-            return `<pre class="hljs"><code>${code}</code></pre>`;
-        },
-    });
-
-    const applyLineNumbers = (code) => {
-        const lines = code.trim().split("\n");
-
-        const rows = lines.map((line, idx) => {
-            const lineNumber = idx + 1;
-
-            let html = "<tr>";
-            html += `<td class="line-number">${lineNumber}</td>`;
-            html += `<td class="code-line">${line}</td>`;
-            html += "</tr>";
-            return html;
-        });
-
-        return `<table><tbody>${rows.join("")}</tbody></table>`;
-    };
 
     onMount(async () => {
         post = await getPost($page.params.slug);
@@ -62,10 +21,9 @@
         setTimeout(() => {
             loading = false;
         }, 200);
+
         title = post.data ? post.data.title : "No blog post found!";
-        content = post.data.content
-            ? md.render(post.data.content)
-            : "No content";
+        content = post.data ? post.data.content : "";
 
         console.log(post);
     });
@@ -78,17 +36,14 @@
 
 {#if loading}
     <img id="loading" src={loadingGif} alt="Loading..." />
-{:else}
+{:else if post.data}
     <div id="blog-post">
         <h1>{title}</h1>
-        {#if post.data}
-            <div>{@html content}</div>
-        {:else}
-            <h2>
-                <NoBlog id={post.id} />
-            </h2>
-        {/if}
+        <MarkDown raw={content} />
     </div>
+{:else}
+    <h1>{title}</h1>
+    <NoBlog id={post.id} />
 {/if}
 
 <style>
